chore(structure): move subscribe modal inside its own folder
