+++ client/src/App.vue
@@ -10,7 +10,7 @@
 <script>
   import AppHeader from './components/AppHeader/AppHeader.vue'
   import FeedbackModal from './components/FeedbackModal.vue'
-  import SubscribeModal from './components/SubscribeModal.vue'
+  import SubscribeModal from './components/SubscribeModal/SubscribeModal.vue'
 
   export default {
     name: 'App',
