+++ client/src/components/SubscribeModal/SubscribeModal.vue
@@ -59,8 +59,8 @@
 </template>
 
 <script>
-  import notificationsService from '../services/notifications'
-  import subscriptionsApi from '../api/subscriptions'
+  import notificationsService from '../../services/notifications'
+  import subscriptionsApi from '../../api/subscriptions'
 
   export default {
     name: 'SubscribeModal',
